Principios Solid
Buenas, de manera colaborativa nos toco realizar una tarea aplicando los principios solid, para ello debediamos buscar repositorios en GitHub y me toque con el suyo. Me tome el atrevimiento de modificar en mi repositorio local:
Simple responsability principle:
Dentro de la clase DatosEspectadores.java tiene funciones en donde realiza otras funcionalidades como la de validaciones de Espectadores, esto podria ser implementado en otra clase.
OpenClose and Liskov, no se encontro incumplimientos con estos principios, pero a un futuro cuando el sistema crezca pueden aparecer estos errores: cuando se quiera implementar mas espectadores o surgan diferentes tipos de espectadores.
