+++ src/java/Datos/DatosEspectadores.java
@@ -10,26 +10,7 @@ import java.util.Hashtable;
 public class DatosEspectadores extends BibliotecaDB
 {
 
-    public Espectador Login(String Usuario, String Password) throws SQLException, ClassNotFoundException, Exception
-    {
-        Password = StringMD.getCadenaEncriptada(Password, StringMD.MD5);
-        String sql = "SELECT * FROM Usuarios WHERE Usuario= '" + Usuario + "' AND Password= '" + Password + "' AND Estado = 1";
-        Conectar();
-        PreparedStatement s = CrearSentencia(sql);
-        ResultSet rows = Consultar(s);        
-        int idEspectador;
-        Espectador aux = null;
-
-        while (rows.next())
-        {
-            idEspectador = rows.getInt("IdUsuario");
-            aux = BuscarEspectador(idEspectador);
-        }
-        
-        Desconectar();
-
-        return aux;
-    }
+    
 
     public void AgregarEspectador(Espectador oEspectador) throws SQLException, ClassNotFoundException, Exception
     {
@@ -66,24 +47,7 @@ public class DatosEspectadores extends BibliotecaDB
         }
     }
 
-    public boolean ComprobarPassword(Espectador oEspectador) throws SQLException, ClassNotFoundException, Exception
-    {
-        boolean rta = false;
-
-        if (oEspectador != null)
-        {
-            String newpass = StringMD.getCadenaEncriptada(oEspectador.getPassword(), StringMD.MD5);
-            Espectador aux = BuscarEspectador(oEspectador.getIdEspectador());
-            String oldpass = aux.getPassword();
-
-            if (!newpass.equals(oldpass))
-            {
-                rta = true;
-            }
-        }
-
-        return rta;
-    }
+    
 
     public void EliminarEspectador(int IdEspectador) throws SQLException, ClassNotFoundException, Exception
     {
