+++ src/java/Datos/Espectador_Mudo.java
@@ -1,0 +1,7 @@
+package Datos;
+
+import Entidades.Espectador;
+
+public class Espectador_Mudo extends Espectador {
+
+}
