+++ src/java/Datos/Espector_Sordo.java
@@ -1,0 +1,7 @@
+package Datos;
+
+import Entidades.Espectador;
+
+public class Espector_Sordo extends Espectador {
+
+}
